new
new

+++ recog_img_crop_process.py
@@ -73,20 +73,21 @@ class Box(object):
 
 
 
-def ProcessEachImgByRotation(srcImgDir, dstImgPath):
+def ProcessEachImgByRotation(srcImgDir, srcImgList, dstImgPath):
 	'''
 	Rotate each img by a certain degree range, save their names and gt
 
 	dstImgPath: location to save the rotated img
 	'''
 
-	gt_file = open(srcImgDir + "filelist.txt")
+	gt_file = open(srcImgDir + srcImgList)
 	gtlines = gt_file.readlines()
 
 	lines = []
 
 	for line in gtlines:
 	    lines.append(line)
+
 
 	rotatedFileNames = []
 	rotatedEquations = []
@@ -95,7 +96,7 @@ def ProcessEachImgByRotation(srcImgDir, dstImgPath):
 	for i in range(len(lines)):
 		curline = lines[ i ].split("\n")[ 0 ]
 
-		curFileName = os.path.join(srcImgDir, curline)
+		curFileName = os.path.join(srcImgDir, curline.split(" ")[ 0 ])
 
 		if not os.path.exists(curFileName):
 			print curFileName + " not exist!"
@@ -103,7 +104,7 @@ def ProcessEachImgByRotation(srcImgDir, dstImgPath):
 
 		print "curFile: " + curFileName
 		img = Image.open(curFileName)
-		print np.shape(img)
+		# print np.shape(img)
 		# img.show()
 
 		img_center_x = np.shape(img)[ 1 ]
@@ -165,7 +166,7 @@ def CheckOverlap(lines):
 			if j in igrLines:
 				continue
 
-			print i, j
+			# print i, j
 
 			if allBox[ i ].iou(allBox[ j ]) > 0:
 				igrLines.append(i)
@@ -177,10 +178,10 @@ def CheckOverlap(lines):
 
 
 
-def CropImgWithAngle(oldcenter, newcenter, angle, ImgName, orgImg, newImg, srcImgDir, dstImgPath):
-	print "Img name: ", ImgName
+def CropImgWithAngle(oldcenter, newcenter, angle, imgName, orgImg, newImg, srcImgDir, dstImgPath):
+	print "Img name: ", imgName
 
-	dstImgList = os.path.join(srcImgDir, os.path.splitext(os.path.basename(ImgName))[0] + ".rec")
+	dstImgList = os.path.join(srcImgDir, os.path.splitext(os.path.basename(imgName))[0] + ".rec")
 	if not os.path.exists(dstImgList):
 		print dstImgList + " not exist!"
 		sys.exit(1)
@@ -195,7 +196,7 @@ def CropImgWithAngle(oldcenter, newcenter, angle, ImgName, orgImg, newImg, srcIm
 	igrLines = []
 	igrLines = CheckOverlap(lines)
 
-	print igrLines
+	# print igrLines
 	for i in range(len(lines)):
 
 		if i in igrLines: # this bbox overlaps with other bboxes, direct ignore 
@@ -256,18 +257,23 @@ def CropImgWithAngle(oldcenter, newcenter, angle, ImgName, orgImg, newImg, srcIm
 		# org_img_crop.show()
 		# img_crop.show()
 
-		print "org_coord: ", x0,y0,x1,y1, "-------  center: ", xold, yold
-		print "new_coord", newxs, newys, newxe, newye, "-------  center: ", xnew, ynew, "-------  shape: ", np.shape(img_crop)
-		img_crop.save(os.path.join(dstImgPath, os.path.splitext(os.path.basename(ImgName))[0] + '_angle_{0}_idx_{1}'.format(angle, i) + ".jpg"))		
-		print "Saving crop img: " + os.path.splitext(os.path.basename(ImgName))[0] + '_angle_{0}'.format(angle) + ".jpg"
+		# print "org_coord: ", x0,y0,x1,y1, "-------  center: ", xold, yold
+		# print "new_coord", newxs, newys, newxe, newye, "-------  center: ", xnew, ynew, "-------  shape: ", np.shape(img_crop)
+		imgFileName = os.path.splitext(os.path.basename(imgName))[0] + '_angle_{0}_idx_{1}'.format(angle, i) + ".jpg"
+
+		img_crop.save(os.path.join(dstImgPath, imgFileName))
+		SaveNewAnnotation(imgFileName, content, dstImgPath)	
+		# print "Saving crop img: " + os.path.splitext(os.path.basename(imgName))[0] + '_angle_{0}'.format(angle) + ".jpg"
 
 
-def SaveNewAnnotation(rotatedFileNames, rotatedEquations, dstImgPath):
+def SaveNewAnnotation(imgFileName, imgLabel, dstImgPath):
 	dstImgList = os.path.join(dstImgPath, "filelist.txt")
-	writer = open(dstImgList, 'w')
 
-	for i in range(len(rotatedFileNames)):
-		writer.write(rotatedFileNames[ i ] + " " + rotatedEquations[ i ] + "\n")
+	if os.path.exists(dstImgList):
+		writer = open(dstImgList, 'w')
+	else:
+		writer = open(dstImgList, 'a')
+	writer.write(imgFileName + " " + imgLabel + "\n")
 
 # def MergeMultipleDataSet():
 
@@ -275,7 +281,8 @@ def SaveNewAnnotation(rotatedFileNames, rotatedEquations, dstImgPath):
 if __name__ == '__main__': 
 
     srcImgDir = sys.argv[1] 
-    dstImgPath = sys.argv[2]
+    srcImgList = sys.argv[2]
+    dstImgPath = sys.argv[3]
 
     if not os.path.exists(srcImgDir):
         print "Input filelist not exists!"
@@ -284,5 +291,5 @@ if __name__ == '__main__':
     if not os.path.exists(dstImgPath):
         os.mkdir(dstImgPath)    
 
-    ProcessEachImgByRotation(srcImgDir, dstImgPath)
-    # SaveNewAnnotation(rotatedFileNames, rotatedEquations, dstImgPath)
+    ProcessEachImgByRotation(srcImgDir, srcImgList, dstImgPath)
+
